+++ include/battle_message.h
@@ -198,13 +198,13 @@ struct BattleMsgData
     u16 currentMove;
     u16 originallyUsedMove;
     u16 lastItem;
-    u8 lastAbility;
+    u16 lastAbility;
     u8 scrActive;
     u8 bakScriptPartyIdx;
     u8 hpScale;
     u8 itemEffectBattler;
     u8 moveType;
-    u8 abilities[4];
+    u16 abilities[4];
     u8 textBuffs[3][0x10];
 };
